requirements.txt
crewai==0.55.2
crewai_tools==0.12.0
fastapi==0.114.0
langchain==0.2.16
langchain_cohere==0.1.9
langchain_community==0.2.16
langchain_mongodb==0.1.9
langchain_openai==0.1.23
pydantic==2.9.1
pymongo==4.7.1
python-dotenv==1.0.1
qdrant_client==1.11.1
torch==2.4.0
transformers==4.44.0
gunicorn==23.0.0
sentence-transformers==3.0.1
Flask-Cors==5.0.0